+++ laktory/models/grants/cataloggrant.py
@@ -1,6 +1,6 @@
 from typing import Literal
 from laktory.models.base import BaseModel
-from laktory.models.grants.databasegrant import PRIVILEGES as DB_PRIVILEGES
+from laktory.models.grants.schemagrant import PRIVILEGES as DB_PRIVILEGES
 
 PRIVILEGES = DB_PRIVILEGES + [
     "USE_CATALOG",