+++ laktory/models/table.py
@@ -26,7 +26,7 @@ class Table(BaseModel):
     primary_key: Union[str, None] = None
     comment: Union[str, None] = None
     catalog_name: Union[str, None] = None
-    database_name: Union[str, None] = None
+    schema_name: Union[str, None] = None
     grants: list[TableGrant] = None
 
     # Data
@@ -50,7 +50,7 @@ class Table(BaseModel):
         for c in self.columns:
             c.table_name = self.name
             c.catalog_name = self.catalog_name
-            c.database_name = self.database_name
+            c.schema_name = self.schema_name
         return self
 
     # ----------------------------------------------------------------------- #
@@ -63,11 +63,11 @@ class Table(BaseModel):
         if self.catalog_name:
             _id += self.catalog_name
 
-        if self.database_name:
+        if self.schema_name:
             if _id == "":
-                _id = self.database_name
+                _id = self.schema_name
             else:
-                _id += f".{self.database_name}"
+                _id += f".{self.schema_name}"
 
         return _id
 
@@ -79,8 +79,8 @@ class Table(BaseModel):
         return _id
 
     @property
-    def schema_name(self) -> str:
-        return self.database_name
+    def database_name(self) -> str:
+        return self.schema_name
 
     # ----------------------------------------------------------------------- #
     # Properties                                                              #
@@ -128,7 +128,7 @@ class Table(BaseModel):
         # Set table
         return Table(
             name="tables",
-            database_name="laktory",
+            schema_name="laktory",
             columns=columns,
         )