+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## [0.0.6]
+## [0.0.6] - 2023-10-09
 **Added**
 - User, service principal and users group models
 - Grants models